Kernel boot improvement
The kernel is now able to boot and place itself  on the right virtual address. Virtual base had to be changed from 0xFFFFFFFFE0000000 to 0xFFFFFFFFC0000000. A bug with the kernel
logging was fixed successfully and we now
can log messages without trouble.
It would be nice to finally add the capacity
to call global constructors.